+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * Update format for file types without description: Connections, Incident Types, Indicator Types, Layouts, Incident Fields.
 * Added a validation for feedTags param in feeds in **validate** command.
 * Fixed readme validation issue in community support packs.
+* Added the **openapi-codegen** command to generate integrations from OpenAPI specification files.
 * Added validation for image links in README files in **validate** command.
 
 # 1.1.7